+++ packages/apollo-codegen-swift/src/codeGeneration.ts
@@ -62,6 +62,7 @@ export interface Options {
 export function generateSource(
   context: CompilerContext,
   outputIndividualFiles: boolean,
+  suppressMultilineStringLiterals: boolean,
   only?: string
 ): SwiftAPIGenerator {
   const generator = new SwiftAPIGenerator(context);
@@ -98,13 +99,21 @@ export function generateSource(
           () => {
             Object.values(context.operations).forEach(operation => {
               if (operation.filePath === inputFilePath) {
-                generator.classDeclarationForOperation(operation, true);
+                generator.classDeclarationForOperation(
+                  operation,
+                  true,
+                  suppressMultilineStringLiterals
+                );
               }
             });
 
             Object.values(context.fragments).forEach(fragment => {
               if (fragment.filePath === inputFilePath) {
-                generator.structDeclarationForFragment(fragment, true);
+                generator.structDeclarationForFragment(
+                  fragment,
+                  true,
+                  suppressMultilineStringLiterals
+                );
               }
             });
           }
@@ -120,11 +129,19 @@ export function generateSource(
       });
 
       Object.values(context.operations).forEach(operation => {
-        generator.classDeclarationForOperation(operation, false);
+        generator.classDeclarationForOperation(
+          operation,
+          false,
+          suppressMultilineStringLiterals
+        );
       });
 
       Object.values(context.fragments).forEach(fragment => {
-        generator.structDeclarationForFragment(fragment, false);
+        generator.structDeclarationForFragment(
+          fragment,
+          false,
+          suppressMultilineStringLiterals
+        );
       });
     });
   }
@@ -158,7 +175,8 @@ export class SwiftAPIGenerator extends SwiftGenerator<CompilerContext> {
    */
   classDeclarationForOperation(
     operation: Operation,
-    outputIndividualFiles: boolean
+    outputIndividualFiles: boolean,
+    suppressMultilineStringLiterals: boolean
   ) {
     const {
       operationName,
@@ -205,10 +223,10 @@ export class SwiftAPIGenerator extends SwiftGenerator<CompilerContext> {
       },
       () => {
         if (source) {
-          this.commentWithoutTrimming(source);
+          this.comment("The raw GraphQL definition of this operation.");
           this.printOnNewline(swift`public let operationDefinition =`);
           this.withIndent(() => {
-            this.multilineString(source);
+            this.multilineString(source, suppressMultilineStringLiterals);
           });
         }
 
@@ -310,7 +328,8 @@ export class SwiftAPIGenerator extends SwiftGenerator<CompilerContext> {
    */
   structDeclarationForFragment(
     { fragmentName, selectionSet, source }: Fragment,
-    outputIndividualFiles: boolean
+    outputIndividualFiles: boolean,
+    suppressMultilineStringLiterals: boolean
   ) {
     const structName = this.helpers.structNameForFragmentName(fragmentName);
 
@@ -323,10 +342,10 @@ export class SwiftAPIGenerator extends SwiftGenerator<CompilerContext> {
       outputIndividualFiles,
       () => {
         if (source) {
-          this.commentWithoutTrimming(source);
+          this.comment("The raw GraphQL definition of this fragment.");
           this.printOnNewline(swift`public static let fragmentDefinition =`);
           this.withIndent(() => {
-            this.multilineString(source);
+            this.multilineString(source, suppressMultilineStringLiterals);
           });
         }
       }